Build request from JsonValue using stream
This is the "proper" way to do it, as it avoid loading a Json string in memory.
The solution is to use our own implementation of PushStreamContent:
https://github.com/aspnet/AspNetWebStack/blob/master/src/System.Net.Http.Formatting/PushStreamContent.cs#L109
The goal is to reduce the need to allocate memory for the request as it is transformed from Json to string to stream etc.

Two new unit tests have also been written in C#, which test that the new method produces the same results as the old method,
even with large amounts of data.