Cleanup and catchup to api
A suggestion for general improvements to existing code in assets/timeseries. The goal is to improve consistency in general, both in the API and internally in the code.

In addition some superfluous code has been removed, and a few missing features have been added to make the current interface match the API docs.