Remove misleading bool return values
Bool as return implies that a failed insertion will result in false. This is not the case for us, in fact, false as a return value is impossible.

This commit changes the return to unit for F# or Task for C#. Empty return implies that there is no return value, and that a failure will result
in either an Error type for F# or an exception for C# like everything else in the SDK.

In UpdateAssets the return value is actually the same as in CreateAssets, the complete list of changed assets, so there bool was changed to that instead.