Gzip datapoints support (#261)
GZip support when writing datapoints. It is a separate method with a "CompressionLevel" parameter.

The simple code used for testing performance improvements is added to the C# playground project. The cutoff is around 10k numeric datapoints, the number would be different and more difficult to pin down for string datapoints. At the maximum chunk size of 100k datapoints the response time is improved 2-3 times.