Groups support (#268)
Implementation for supporting groups. I've been looking into this because I'd like to list capabilities in token/inspect. A custom converter is necessary no matter what for properly serializing and deserializing capabilities. In order to make the user interface pleasant, and not plagued by complicated nested objects, I've made a custom converter which takes takes objects on the form

```
class SpecificCapability {
    IEnumerable<string> Actions
    ScopeType1 SomeScope
    ScopeType2 SomeOtherScope
}
```

and converts them to

```
{
    "specificCapability": {
        "actions": [...],
        "scope": {
            "someScope": { ... }
        }
    }
}
```

This required a decent amount of custom converter code.

I'm very open for suggestions for how to do this better, but I've tried multiple approaches, and this gave the best results in my opinion.