Fix bug serializing ACLs with idscope (#276)
Turns out docs are wrong here, and there is probably something funky with how this is done in the auth API itself. IdScope is documented to be (mostly) arrays of integers, but they are universally arrays of strings, although the API accepts arrays of integers on groups create.

We solve this by just using integers locally (which is better for the user), and sending integers to the API (since it works, and is technically more compact). There is a convenient attribute in System.Text.Json which handles this for us.

I also added a few tests, to make sure this actually works.