Transformations (#282)
Adds transformations support to the SDK.

This is a pretty big commit, since transformations is a pretty complicated resource type. There isn't much that is anything special, except for the deserialization of table schemas, which required a small custom converter, since there is a type that is effectively union of string and object.