Templates (#286)
I'm doing a project with templates and need proper support. This PR implements that, more or less. There are a few places with JsonElement, which is due to a mix of lacking documentation and technical challenges (the templates API is not documented, so this is based on the source code and the openAPI spec). Since it's still in beta I guess we can make breaking changes without having to get a new major version in the SDK.