Dms v3 (#307)
All of DMS V3, except a couple details I couldn't figure out. This replaces the old V2 implementation. I don't think it's necessary that we support both.

DMS actually contains all three types of union representations (untagged, internally tagged, and externally tagged), and is in general quite complicated, so this implementation contains a large number of custom json converters, and three base converters for handling unions. They might come in handy later as well.

There are tests, they are a little simple, and probably should be expanded in the future.

The requests to CDF end up being _very_ clunky to write, and involving a lot of lookup of subtypes. I've documented this in some places, but more work making it more ergonomic could absolutely be done.

I put everything into a separate namespace, `Beta.DataModels`. Types are mostly in the base namespace, but the number of types makes that impractical in this case.